Split out Alexey's matrix code into its own module
--HG--
rename : Statistics/Test/KolmogorovSmirnov.hs => Statistics/Matrix.hs
